extract use paste hook
